Refactor: login, token 파일 분리해서 정리
